+++ HISTORY.md
@@ -8,6 +8,8 @@
   ([#463](https://github.com/python-attrs/cattrs/pull/463))
 - More robust support for `Annotated` and `NotRequired` in TypedDicts.
   ([#450](https://github.com/python-attrs/cattrs/pull/450))
+- `typing_extensions.Literal` is now automatically structured, just like `typing.Literal`.
+  ([#460](https://github.com/python-attrs/cattrs/issues/460) [#467](https://github.com/python-attrs/cattrs/pull/467))
 - [PEP 695](https://peps.python.org/pep-0695/) generics are now tested.
   ([#452](https://github.com/python-attrs/cattrs/pull/452))
 - Imports are now sorted using Ruff.